+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:60648/","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:60628/","mongoDBName":"jest"}
