+++ src/data/usecases/add-account/db-add-account-protocols.ts
@@ -1,4 +1,4 @@
 export * from '@domain/models/account'
 export * from '@domain/usecases/add-account'
 export * from '@data/protocols/cryptography/hasher'
-export * from '@data/protocols/db/add-account-repository'
+export * from '@data/protocols/db/account/add-account-repository'
