+++ src/infrastructure/db/mongodb/account/account-mongo-repository.spec.ts
@@ -1,6 +1,6 @@
 import { Collection, ObjectId } from 'mongodb';
 import { MongoHelper } from '../helpers/mongo-helper'
-import { AccountMongoRepository } from './account';
+import { AccountMongoRepository } from './account-mongo-repository';
 
 let accountCollection: Collection
