+++ src/infrastructure/db/mongodb/log/log-mongo-repository.ts
@@ -1,5 +1,5 @@
 
-import { LogErrorRepository } from '@data/protocols/db/log-error-repository'
+import { LogErrorRepository } from '@data/protocols/db/log/log-error-repository'
 import { MongoHelper } from '../helpers/mongo-helper'
 
 export class LogMongoRepository implements LogErrorRepository {
