+++ src/main/decorators/log-controller-decorator.spec.ts
@@ -1,8 +1,8 @@
 import { Controller, HttpRequest, httpResponse } from '@presentation/protocols'
-import { LogControllerDecorator } from './log'
+import { LogControllerDecorator } from './log-controller-decorator'
 import { serverError, ok } from '@presentation/helpers/http/http-helper'
 import { AccountModel } from '@domain/models/account'
-import { LogErrorRepository } from '@data/protocols/db/log-error-repository'
+import { LogErrorRepository } from '@data/protocols/db/log/log-error-repository'
 
 const makeController = (): Controller => {
   class ControllerStub implements Controller {
