+++ src/main/decorators/log-controller-decorator.ts
@@ -1,4 +1,4 @@
-import { LogErrorRepository } from '@data/protocols/db/log-error-repository'
+import { LogErrorRepository } from '@data/protocols/db/log/log-error-repository'
 import { Controller, HttpRequest, httpResponse } from '@presentation/protocols'
 
 export class LogControllerDecorator implements Controller {
