+++ src/main/factories/login/login-validation-factory.spec.ts
@@ -1,7 +1,7 @@
 import { RequiredFieldValidation, EmailValidation, ValidationComposite } from '@presentation/helpers/validators'
 import { EmailValidator } from '@presentation/protocols/email-validator.interface'
 import { Validation } from '@presentation/protocols/validation'
-import { makeLoginValidation } from './login-validation'
+import { makeLoginValidation } from './login-validation-factory'
 
 jest.mock('@presentation/helpers/validators/validation-composite')
