+++ src/main/factories/signup/index.ts
@@ -1,11 +1,11 @@
 import { DbAddAccount } from '@data/usecases/add-account/db-add-account'
 import { BcryptAdapter } from '@infrastructure/cryptography/bcrypt-adapter/bcrypt-adapter'
-import { AccountMongoRepository } from '@infrastructure/db/mongodb/account-repository/account'
-import { LogMongoRepository } from '@infrastructure/db/mongodb/log-repository/log'
-import { LogControllerDecorator } from '@main/decorators/log'
-import { SignUpController } from '@presentation/controller/signup/signup'
+import { AccountMongoRepository } from '@infrastructure/db/mongodb/account/account-mongo-repository'
+import { LogMongoRepository } from '@infrastructure/db/mongodb/log/log-mongo-repository'
+import { LogControllerDecorator } from '@main/decorators/log-controller-decorator'
+import { SignUpController } from '@presentation/controller/signup/signup-controller'
 import { Controller } from '@presentation/protocols'
-import { makeSignUpValidation } from './signup-validation'
+import { makeSignUpValidation } from './signup-validation-factory'
 
 export const makeSignUpController = (): Controller => {
   const salt = 12
