+++ src/main/factories/signup/signup-validation-factory.spec.ts
@@ -6,7 +6,7 @@ import {
 } from '@presentation/helpers/validators'
 import { EmailValidator } from '@presentation/protocols/email-validator.interface'
 import { Validation } from '@presentation/protocols/validation'
-import { makeSignUpValidation } from './signup-validation'
+import { makeSignUpValidation } from './signup-validation-factory'
 
 jest.mock('@presentation/helpers/validators/validation-composite')
