+++ src/presentation/controller/login/login-controller.spec.ts
@@ -1,7 +1,7 @@
 import { MissingParamError } from '@presentation/errors'
 import { badRequest, ok, serverError, unauthorized } from '@presentation/helpers/http/http-helper'
-import { LoginController } from './login'
-import { Authentication, AuthenticationModel, HttpRequest, Validation } from './login-protocols'
+import { LoginController } from './login-controller'
+import { Authentication, AuthenticationModel, HttpRequest, Validation } from './login-controller-protocols'
 
 const makeAuthentication = (): Authentication => {
   class AuthenticationStub implements Authentication {
