+++ src/presentation/controller/signup/signup-controller.spec.ts
@@ -1,7 +1,7 @@
 import { MissingParamError, ServerError } from '@presentation/errors'
 import { badRequest, ok, serverError } from '@presentation/helpers/http/http-helper'
-import { SignUpController } from './signup'
-import { AccountModel, AddAccount, AddAccountModel, HttpRequest, Validation } from './signup-protocols'
+import { SignUpController } from './signup-controller'
+import { AccountModel, AddAccount, AddAccountModel, HttpRequest, Validation } from './signup-controller-protocols'
 
 const makeAddAccount = (): AddAccount => {
   // factory
